sho/requirements.txt
# sho/requirements.txt — SHO x402 Proxy dependencies

# Web framework (same as SLO oracle backends)
fastapi>=0.104.0
uvicorn>=0.24.0

# HTTP client for backend calls and Base RPC
httpx>=0.25.0

# Ed25519 signing
PyNaCl>=1.5.0

# secp256k1 (for cross-certification — already installed for SLO)
ecdsa>=0.18.0
